% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_predictions.R
\name{plot_prediction_explanation}
\alias{plot_prediction_explanation}
\title{Creates a breakdown plot for prediction using Sequential Variable Attributions Conditioning}
\usage{
plot_prediction_explanation(model, explainer, newdata, order = NULL, ...)
}
\arguments{
\item{model}{a fitted model}

\item{explainer}{a DALEX explainer object for the model}

\item{newdata}{a data frame for which to calculate predictions}

\item{order}{Order of variables, passed to iBreakDown::break_down function}
}
\value{
a breakdown plot for the calculated prediction
}
\description{
Creates a breakdown plot for prediction using Sequential Variable Attributions Conditioning
}
\author{
Jacek Podlewski
}
