% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_predictions.R
\name{plot_prediction_profile}
\alias{plot_prediction_profile}
\title{Creates a Ceteris Paribus profile plot for prediciton}
\usage{
plot_prediction_profile(model, explainer, new_data, grid = "quantile", ...)
}
\arguments{
\item{model}{a fitted model}

\item{explainer}{a DALEX explainer object for the model}

\item{grid}{type of point grid to use on the graph, 'quantiles' or 'uniform'}

\item{newdata}{a data frame for which to calculate predictions}
}
\value{
a Ceteris Paribus profile plot for the calculated prediction
}
\description{
Creates a Ceteris Paribus profile plot for prediciton
}
\author{
Jacek Podlewski
}
