% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_predictions.R
\name{plot_shap}
\alias{plot_shap}
\title{Creates a SHAP plot for prediciton}
\usage{
plot_shap(model, explainer, new_data, max_features = 5, ...)
}
\arguments{
\item{model}{a fitted model}

\item{explainer}{a DALEX explainer object for the model}

\item{max_features}{maximal number of features to show on plot}

\item{newdata}{a data frame for which to calculate predictions}
}
\value{
a SHAP plot for the calculated prediction
}
\description{
Creates a SHAP plot for prediciton
}
\author{
Jacek Podlewski
}
