% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_predictions.R
\name{prediction_probs}
\alias{prediction_probs}
\title{Calculates positive SPECT-CT result probability from model}
\usage{
prediction_probs(model, newdata)
}
\arguments{
\item{model}{a fitted model}

\item{newdata}{a data frame for which to calculate predictions}
}
\value{
a dataframe with predicted probability converted to % in 0-100 range and predicted result (Positive/Negative)
}
\description{
Calculates positive SPECT-CT result probability from model
}
\author{
Jacek Podlewski
}
